Optimistically updating document after update
